+++ frontend/src/metabase/common/data-studio/selectors.ts
@@ -3,7 +3,7 @@ import type { State } from "metabase/redux/store";
 import { getIsEmbeddingIframe } from "metabase/selectors/embed";
 import { getUserIsAdmin, getUserIsAnalyst } from "metabase/selectors/user";
 
-// Must be in sync with CanAccessDataStudio in frontend/src/metabase/router/guards.tsx
+// Must be in sync with CanAccessDataStudio in frontend/src/metabase/data-studio/route-guards.tsx
 export function canAccessDataStudio(state: State) {
   if (getIsEmbeddingIframe(state)) {
     return false;
