+++ frontend/src/metabase/data-studio/routes.tsx
@@ -27,6 +27,7 @@ import { TransformsSectionLayout } from "./app/pages/TransformsSectionLayout";
 import { WorkspacesSectionLayout } from "./app/pages/WorkspacesSectionLayout";
 import { getDataStudioMetadataRoutes } from "./data-model/routes";
 import { getDataStudioGlossaryRoutes } from "./glossary/routes";
+import { CanAccessDataModel, CanAccessDataStudio } from "./route-guards";
 import { getDataStudioSettingsRoutes } from "./settings/routes";
 import {
   DependenciesUpsellPage,
@@ -36,11 +37,7 @@ import {
 
 const RoutedTransformsSectionLayout = withRouteProps(TransformsSectionLayout);
 
-export function getDataStudioRoutes(
-  CanAccessDataStudio: RouteComponent,
-  CanAccessDataModel: RouteComponent,
-  IsAdmin: RouteComponent,
-) {
+export function getDataStudioRoutes(IsAdmin: RouteComponent) {
   return (
     <>
       {/* These redirects sit
