+++ frontend/src/metabase/route-guards/index.ts
@@ -1,0 +1,11 @@
+export {
+  AvailableInEmbedding,
+  CanAccessOnboarding,
+  CanAccessSettings,
+  IsAdmin,
+  IsAuthenticated,
+  IsNotAuthenticated,
+  MetabaseIsSetup,
+  UserIsAuthenticated,
+} from "./auth-guards";
+export { createGuard, createRedirectGuard } from "./create-guard";
